run tests before auto committing.

+++ tests/run_test.sh
@@ -13,19 +13,20 @@ test -d "${VK_HEADERS_SANDBOX}" && (echo "run gen_vku.sh first." 1>&2 && false)
 
 if [[ "${1}" == "--dev" ]]; then
   VK_VERSION="v0.0.0"
-else
+fi
+
+if [[ ! -n "${VK_VERSION}" ]]; then
   VK_HEADERS_TAG=$(
     cd "${VK_HEADERS_SANDBOX}" &&
     git describe --tag | grep -e "${VK_SDK_VER//\./\\.}" | sort -Vr | head -1
     )
-
   VK_VERSION="${VK_HEADERS_TAG//[^0-9.]/}"
   # prepend v and strip the fix number. most recent fix version should always prevail.
   VK_VERSION=$(set ${VK_VERSION//\./ }; echo "v${1}.${2}.${3}")
+fi
 
-  if [[ -d build ]]; then
-    /bin/rm -rf build
-  fi
+if [[ "${VK_VERSION}" != "v0.0.0" && -d build ]]; then
+  /bin/rm -rf build
 fi
 
 set -eux
